Rename ContentInstallInfo -> InstallInfo
