+++ ansible_galaxy/models/install_info.py
@@ -7,7 +7,7 @@ log = logging.getLogger(__name__)
 
 
 @attr.s(frozen=True)
-class ContentInstallInfo(object):
+class InstallInfo(object):
     '''The info that is saved into the .galaxy_install_info file'''
     install_date = attr.ib()
     install_date_iso = attr.ib()
